Remove decrepit and unavaliable API and fix signatures

+++ openssl-sys/src/ssl3.rs
@@ -1,5 +1,5 @@
 use libc::*;
 
-pub const SSL3_VERSION: c_int = 0x300;
+pub const SSL3_VERSION: u16 = 0x300;
 
 pub const SSL3_AD_ILLEGAL_PARAMETER: c_int = 47;
