+++ openssl/src/base64.rs
@@ -15,7 +15,7 @@ use libc::c_int;
 /// [`EVP_EncodeBlock`]: https://www.openssl.org/docs/man1.1.1/man3/EVP_DecodeBlock.html
 pub fn encode_block(src: &[u8]) -> String {
     assert!(src.len() <= c_int::max_value() as usize);
-    let src_len = src.len() as c_int;
+    let src_len = src.len();
 
     let len = encoded_len(src_len).unwrap();
     let mut out = Vec::with_capacity(len as usize);
@@ -49,7 +49,7 @@ pub fn decode_block(src: &str) -> Result<Vec<u8>, ErrorStack> {
     }
 
     assert!(src.len() <= c_int::max_value() as usize);
-    let src_len = src.len() as c_int;
+    let src_len = src.len();
 
     let len = decoded_len(src_len).unwrap();
     let mut out = Vec::with_capacity(len as usize);
@@ -78,7 +78,7 @@ pub fn decode_block(src: &str) -> Result<Vec<u8>, ErrorStack> {
     Ok(out)
 }
 
-fn encoded_len(src_len: c_int) -> Option<c_int> {
+fn encoded_len(src_len: usize) -> Option<usize> {
     let mut len = (src_len / 3).checked_mul(4)?;
 
     if src_len % 3 != 0 {
@@ -90,7 +90,7 @@ fn encoded_len(src_len: c_int) -> Option<c_int> {
     Some(len)
 }
 
-fn decoded_len(src_len: c_int) -> Option<c_int> {
+fn decoded_len(src_len: usize) -> Option<usize> {
     let mut len = (src_len / 4).checked_mul(3)?;
 
     if src_len % 4 != 0 {
